Created playerStats View
It's ok to break things from big pieces down to smaller ones. You have
to have info on a table to work with it. Joining information from
players table (which will already have entries) to the matches table
with a left join (which allow for all columns that were not matched with
players table to be shown opened up a flood gate of discovery and "Ah
ha" moments was able to get playerWins view and matchesPlayed view to
work pretty well after that...REMEMBER BIG to small when working with
these tables.